+++ packages/expo-router/src/global-state/routing.ts
@@ -1,6 +1,7 @@
 import { StackActions, type NavigationState, PartialRoute } from '@react-navigation/native';
 import * as Linking from 'expo-linking';
 import { nanoid } from 'nanoid/non-secure';
+import { Platform } from 'react-native';
 
 import { type RouterStore } from './router-store';
 import { ResultState } from '../fork/getStateFromPath';
@@ -101,6 +102,10 @@ export function linkTo(
   { event, relativeToDirectory, withAnchor }: LinkToOptions = {}
 ) {
   if (shouldLinkExternally(href)) {
+    if (href.startsWith('//') && Platform.OS !== 'web') {
+      href = `https:${href}`;
+    }
+
     Linking.openURL(href);
     return;
   }